find house by id

+++ src/main/java/rental/infrastructure/persistence/HouseJpaPersistence.java
@@ -3,5 +3,5 @@ package rental.infrastructure.persistence;
 import org.springframework.data.jpa.repository.JpaRepository;
 import rental.infrastructure.dataentity.HouseEntity;
 
-public interface HouseJpaPersistence extends JpaRepository<HouseEntity, Integer> {
+public interface HouseJpaPersistence extends JpaRepository<HouseEntity, Long> {
 }
